ADM-309 refactor: edit url to restful url and move files
